Githubmergetest

+++ beatsPM/src/main/java/com/example/beatsPM/Models/Data/CommentsRepository.java
@@ -9,7 +9,6 @@ import javax.transaction.Transactional;
 
 @Repository
 @Transactional
-public interface CommentsRepository extends CrudRepository <CommentModel, Integer> {
-
+public interface CommentsRepository extends CrudRepository <CommentModel, Long> {
 
 }
